+++ src/parsers/contest/HUSTOJSyzojContestParser.ts
@@ -1,0 +1,5 @@
+import { SimpleContestParser } from '../SimpleContestParser';
+
+export abstract class HUSTOJSyzojContestParser extends SimpleContestParser {
+  protected linkSelector = 'table.ui.selectable > tbody > tr > td:nth-child(2) > a';
+}